requirements.txt
langchain
faiss-cpu
pypdf
python-docx
google-generativeai
sentence-transformers
gradio
langchain-community
